Classificação MVC
Foram criados pacotes de modelo, controle e view, além da separação do
código nessas pastas.
