+++ Trabalhos/bin/.gitignore
@@ -1,1 +1,4 @@
 /TP04/
+/modelo/
+/controle/
+/view/
