+++ Trabalhos/src/controle/Grupo.java
@@ -1,4 +1,4 @@
-package TP04;
+package controle;
 
 public abstract class Grupo {
