+++ Trabalhos/src/controle/Temperatura.java
@@ -1,4 +1,4 @@
-package TP04;
+package controle;
 
 public class Temperatura {
 	private double valorOrigem;
